requirements.txt
celery==5.3.1
django==4.2.4
django-cors-headers==4.2.0
django_celery_beat==2.5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.2.1
marshmallow==3.19.0
marshmallow_dataclass==8.5.14
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dotenv==1.0.0
redis==5.0.0
requests==2.28.2
telebot==0.0.5


